+++ packages/admin/dashboard/src/hooks/use-combobox-data.tsx
@@ -21,7 +21,7 @@ type ComboboxQueryParams = {
 
 export const useComboboxData = <
   TResponse extends ComboboxExternalData,
-  TParams extends ComboboxQueryParams,
+  TParams extends ComboboxQueryParams
 >({
   queryKey,
   queryFn,