+++ packages/admin/dashboard/src/routes/products/product-attributes/components/product-attributes-form/product-attributes-form.tsx
@@ -74,7 +74,7 @@ export const ProductAttributesForm = ({
         height: data.height ? data.height : null,
         mid_code: data.mid_code,
         hs_code: data.hs_code,
-        origin_country: data.origin_country,
+        origin_country: data.origin_country || null,
       },
       {
         onSuccess: () => {
@@ -246,9 +246,11 @@ export const ProductAttributesForm = ({
                 render={({ field }) => {
                   return (
                     <Form.Item>
-                      <Form.Label>{t("fields.countryOfOrigin")}</Form.Label>
+                      <Form.Label optional>
+                        {t("fields.countryOfOrigin")}
+                      </Form.Label>
                       <Form.Control>
-                        <CountrySelect {...field} />
+                        <CountrySelect {...field} allowClear />
                       </Form.Control>
                       <Form.ErrorMessage />
                     </Form.Item>