+++ packages/admin/dashboard/src/routes/tax-regions/tax-region-tax-override-create/components/tax-region-override-create-form/tax-region-tax-override-create.tsx
@@ -636,6 +636,7 @@ export const TaxRegionCreateTaxOverrideForm = ({
                                               key={field.id}
                                               index={index}
                                               label={field.label}
+                                              value={field.value}
                                               onRemove={remove}
                                             />
                                           )