+++ packages/core/core-flows/src/draft-order/workflows/refresh-draft-order-adjustments.ts
@@ -1,5 +1,6 @@
 import { PromotionActions } from "@medusajs/framework/utils"
 import {
+  createHook,
   createWorkflow,
   parallelize,
   WorkflowData,
@@ -11,6 +12,7 @@ import {
   getPromotionCodesToApply,
   prepareAdjustmentsFromPromotionActionsStep,
 } from "../../cart"
+import { promotionContextResult } from "../../cart/utils/schemas"
 import { createDraftOrderLineItemAdjustmentsStep } from "../steps/create-draft-order-line-item-adjustments"
 import { createDraftOrderShippingMethodAdjustmentsStep } from "../steps/create-draft-order-shipping-method-adjustments"
 import { removeDraftOrderLineItemAdjustmentsStep } from "../steps/remove-draft-order-line-item-adjustments"
@@ -79,6 +81,8 @@ export interface RefreshDraftOrderAdjustmentsWorkflowInput {
  * @summary
  *
  * Refresh the promotions in a draft order.
+ *
+ * @property hooks.setPromotionContext - This hook is executed before promotion rules are evaluated for the draft order. You can consume this hook to return any custom context that should be merged on top of the order context when evaluating promotion rules (e.g. `company.id`, `custom_tier`).
  */
 export const refreshDraftOrderAdjustmentsWorkflow = createWorkflow(
   refreshDraftOrderAdjustmentsWorkflowId,
@@ -89,6 +93,19 @@ export const refreshDraftOrderAdjustmentsWorkflow = createWorkflow(
       ttl: 10,
     })
 
+    const setPromotionContext = createHook(
+      "setPromotionContext",
+      {
+        order: input.order,
+        promo_codes: input.promo_codes,
+        action: input.action,
+      },
+      {
+        resultValidator: promotionContextResult,
+      }
+    )
+    const setPromotionContextResult = setPromotionContext.getResult()
+
     const promotionCodesToApply = getPromotionCodesToApply({
       cart: input.order,
       promo_codes: input.promo_codes,
@@ -98,6 +115,7 @@ export const refreshDraftOrderAdjustmentsWorkflow = createWorkflow(
     const actions = getActionsToComputeFromPromotionsStep({
       computeActionContext: input.order as any,
       promotionCodesToApply,
+      additional_promotion_context: setPromotionContextResult,
     })
 
     const {
@@ -134,6 +152,8 @@ export const refreshDraftOrderAdjustmentsWorkflow = createWorkflow(
       key: input.order.id,
     })
 
-    return new WorkflowResponse(void 0)
+    return new WorkflowResponse(void 0, {
+      hooks: [setPromotionContext] as const,
+    })
   }
 )